requirements/base.txt
Django==4.0.4

django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-filter==22.1
django-import-export==2.8.0
fcm-django==1.0.11
firebase-admin==5.2.0
django-cors-headers==3.12.0
django-redis==5.2.0
celery==5.2.7
django-celery-beat==2.4.0
django-celery-results==2.4.0
